Placed SpecularData in a separate compilation unit
